// src/api/generated/models/stedHistoryResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { StedAssessmentResponseDto } from './stedAssessmentResponseDto';

export interface StedHistoryResponseDto {
  childId: string;
  items: StedAssessmentResponseDto[];
  total: number;
  page?: number;
  pageSize?: number;
  totalPages?: number;
}